refactor(tests): split declaration.test.ts into focused files (< 800 lines each)
